Multivariate optimal allocation

Multivariate optimal allocation for different domains of interest in stratified sample design

bethel  (
          stratif,
          errors,
          minnumstrat=2,
          maxiter=200,
          maxiter1=25,
          printa=FALSE,
                  realAllocation=FALSE,
          epsilon=1e-11
          )

Arguments

errors
Data frame of coefficients of variation for each domain

stratif
Data frame of survey strata

minnumstrat
Minimum number of units per strata (default=2)

maxiter
Maximum number of iterations of the algorithm (default=200)

maxiter1
Maximum number of iterations (default=25) of the general procedure. This kind of iteration may be required by the fact that when in a stratum the number of allocated units is greater or equal to its population, that stratum is set as "census stratum", and the whole procedure is re-initialised

printa
If TRUE then two attributes are added to the resulting vector. The first ('confr') is a comparison between results obtained with 3 different allocation methods: Bethel, proportional and equal. The second ('outcv') is a table reporting planned and actual CV, together with a sensitivity analysis

realAllocation
If FALSE, the allocation is based on INTEGER values; if TRUE, the allocation is based on REAL values

epsilon
Epsilon (default=1e-11)): this value is used to compare the difference in results from one iteration to the other; if it it is lower than "epsilon", then the procedure stops

Value

A vector containing the computed optimal allocation

Examples

library(SamplingStrata)
data(strata)
data(errors)
n <- bethel(strata, errors, printa=TRUE)
sum(n)

#> [1] 168

attributes(n)$confr

#>       STRATUM POPULATION BETHEL PROPORTIONAL EQUAL
#>  [1,] "STR1"  "2246"     "15"   "13"         "14" 
#>  [2,] "STR2"  "2972"     "17"   "17"         "14" 
#>  [3,] "STR3"  "1905"     "11"   "11"         "14" 
#>  [4,] "STR4"  "3125"     "18"   "18"         "14" 
#>  [5,] "STR5"  "1733"     "11"   "10"         "14" 
#>  [6,] "STR6"  "1060"     "4"    "6"          "14" 
#>  [7,] "STR7"  "2928"     "11"   "16"         "14" 
#>  [8,] "STR8"  "2498"     "13"   "14"         "14" 
#>  [9,] "STR9"  "3715"     "16"   "21"         "14" 
#> [10,] "STR10" "3128"     "17"   "18"         "14" 
#> [11,] "STR11" "3868"     "25"   "22"         "14" 
#> [12,] "STR12" "1641"     "10"   "9"          "14" 
#> [13,] "TOTAL" "30819"    "168"  "175"        "168"

attributes(n)$outcv

#>      TYPE   DOMAIN/VAR. PLANNED CV  ACTUAL CV SENSITIVITY 10%
#> [1,] "DOM1" "tot/V1"    "0.04"      "0.0392"  "33"           
#> [2,] "DOM1" "tot/V2"    "0.06"      "0.0332"  "1"