Function that allows to calculate a heteroscedasticity index, together with associate prediction variance, to be used by the optimization step to correctly evaluate the standard deviation in the strata due to prediction errors.

When the anticipated variance has to be calculated during the execution of the optimization step, his function allows to calculate a heteroscedasticity index, together with associate prediction variance, to be used to correctly evaluate the variance in the strata. The function returns a list where the first object is the heteroscedasticity indexand the second is the associated standard deviation in the strata due to prediction errors. These two values can be passed as parameters of the model, or used to calculate prediction errors for ach unit in the frame.

computeGamma(e,x,nbins)

Arguments

e
This is the variable that contains prediction errors (residuals) of the model.

x
This is the variable that contains explanatory variable in the model.

nbins
Number of bins to be passed to the 'var.bin' function. Default is 6.

Value

A list containing: (i) the value of the heteroscedasticity index, (ii) associated standard deviation.

Examples

# NOT RUN {
data("swissmunicipalities")
swiss_sample <- swissmunicipalities[sample(c(1:nrow(swissmunicipalities)),500),]
mod_Airind_POPTOT <- lm(swiss_sample$Airind ~ swiss_sample$POPTOT)
computeGamma(mod_Airind_POPTOT$residuals,
             swiss_sample$POPTOT,
             nbins = 10)
# [1] 0.78830779 0.01418615
# }