Precision constraints (maximum CVs) as input for Bethel allocation

Dataframe containing precision levels (expressed in terms of acceptable CV's)

data(errors)

Format

The constraint data frame (swisserrors) contains a row per each domain value with the following variables:

DOM

Type of domain code (factor)

CV1

Planned coefficient of variation for first variable Y1 (number of men and women aged between 0 and 19) (numeric)

CV2

Planned coefficient of variation for second variable Y2 (number of men and women aged between 20 and 39) (numeric)

CV3

Planned coefficient of variation for third variable Y3 (number of men and women aged between 40 and 64) (numeric)

CV4

Planned coefficient of variation for forth variable Y4 (number of men and women aged between 65 and over) (numeric)

domainvalue

Value of the domain to which the constraints refer (numeric)

Examples

## data(swisserrors)
## swisserrors