Dataframe containing information on strata in the swiss municipalities frame

Dataframe containing information on strata in the swiss municipalities frame

data(swissframe)

Format

The "swissstrata" dataframe contains a row per stratum with the following variables:

STRATO

Identifier of the stratum (character)

N

Number of population units in the stratum (numeric)

X1

Classes of total population in the municipality (factor with 18 values)

X2

Classes of wood area in the municipality (factor with 3 values)

X3

Classes of area under cultivation in the municipality (factor with 3 values)

X4

Classes of mountain pasture area in the municipality (factor with 3 values)

X5

Classes of area with buildings in the municipality (factor with 3 values)

X6

Classes of industrial area in the municipality (factor with 3 values)

M1

Mean in the stratum of Y1 (number of men and women aged between 0 and 19)(numeric)

M2

Mean in the stratum of Y2 (number of men and women aged between 20 and 39) (numeric)

M3

Mean in the stratum of Y3 (number of men and women aged between 40 and 64) (numeric)

M4

Mean in the stratum of Y4 (number of men and women aged between 64 and over) (numeric)

S1

Standard deviation in the stratum of Y1 (number of men and women aged between 0 and 19)(numeric)

S2

Standard deviation in the stratum of Y2 (number of men and women aged between 20 and 39) (numeric)

S3

Standard deviation in the stratum of Y3 (number of men and women aged between 40 and 64) (numeric)

S4

Standard deviation in the stratum of Y4 (number of men and women aged between 64 and over) (numeric)

cens

Flag (1 indicates a take all straum, 0 a sampling stratum) (numeric) Default = 0

cost

Cost per interview in each stratum. Default = 1 (numeric)

DOM1

Value of domain to which the stratum belongs Default = 1 (factor or numeric)

Examples

# data(swissstrata)
# head(swissstrata)