Expected coefficients of variation of target variables Y

Once optimized the sampling frame, this function allows to calculate the expected coefficients of variation on the aggregated strata of the current optimized solution, and to compare them to the precision constraints.

expected_CV(strata)

Arguments

strata

Aggregated strata in the solution obtained by the execution of the 'optimizedStrata' function

Value

Matrix containing values of the CV's in the different domains

Examples

if (FALSE) {
library(SamplingStrata)
data(swisserrors)
data(swissstrata)
# optimisation of sampling strata
solution <- optimizeStrata ( 
    errors = swisserrors, 
    strata = swissstrata, 
)
# calculate CV's on Y's
expected_CV(solution$aggr_strata)
# compare to precision constraints
swisserrors
}