Dataframe containing information on strata in the frame

Dataframe containing information on strata in the frame

data(strata)

Format

The strata data frame (strata) contains a row per stratum with the following variables:

stratum

Identifier of the stratum (numeric)

N

Number of population units in the stratum (numeric)

X1

Value of first auxiliary variable X1 in the stratum (factor)

Xi

Value of i-th auxiliary variable Xi in the stratum (factor)

Xk

Value of last auxiliary variable Xk in the stratum (factor)

M1

Mean in the stratum of the first variable Y1 (numeric)

Mj

Mean in the stratum of the j-th variable Yt (numeric)

Mn

Mean in the stratum of the last variable Y (numeric)

S1

Standard deviation in the stratum of the first variable Y (numeric)

Sj

Standard deviation in the stratum of the j-th variable Yt (numeric)

Sn

Standard deviation in the stratum of the last variable Y (numeric)

cens

Flag (1 indicates a take all straum, 0 a sampling stratum) (numeric) Default = 0

cost

Cost per interview in each stratum. Default = 1 (numeric)

DOM1

Value of domain to which the stratum belongs (factor or numeric)

Details

Note: the names of the variables must be the ones indicated above

Examples

# data(strata)
# head(strata)